move all storage-based functionality into single folder
