+++ libs/site-base/src/containers/AddAccount/Storage/Restore/RestoreChoice.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { FormattedMessage, MessageDescriptor } from 'react-intl';
+import { Link } from 'react-router-dom';
+import { Header } from 'semantic-ui-react';
+
+type Props = {
+  link: string,
+  imgSrc: string,
+  txt: MessageDescriptor
+}
+
+export const RestoreChoice = ({link, imgSrc, txt}: Props) =>
+  <Link to={link}>
+    <img src={imgSrc} />
+    <Header as={"h4"}><FormattedMessage {...txt} /></Header>
+  </Link>
